Split ORANGE testing into subtests
